separate out dash components
