+++ src/components/__init__.py
@@ -1,2 +1,10 @@
 # pull in components from files in the current directory to make imports cleaner
 from .move_log import move_log
+from .reset_graph import reset_graph
+from .algorithms import algorithms
+from .hover_data import hover_data
+from .zoom_data import zoom_data
+from .load_graph import load_graph
+from .measurementbasis import measurementbasis
+from .figure import display_options, figure
+from .error_channel import error_channel
